feat: add authentication labels to locales, improve auth error handling, and refactor refresh token schema migration
