+++ apps/backend/src/db/schema/index.ts
@@ -1,4 +1,4 @@
 export * from "./projects";
 export * from "./users";
 export * from "./ventures";
-export * from "./auth";
+export * from "./refresh-tokens";
