+++ apps/mobile/src/i18n/locales/en.json
@@ -34,7 +34,11 @@
     "whatsapp_placeholder": "3624 000000",
     "first_name_label": "First Name",
     "last_name_label": "Last Name",
-    "submit_button": "Get Started"
+    "submit_button": "Get Started",
+    "email_label": "Email",
+    "password_label": "Password",
+    "admin_portal": "Administrator Portal",
+    "entrepreneur_panel": "Entrepreneur Panel"
   },
   "validation": {
     "name_required": "Name is required",
@@ -218,7 +222,8 @@
     "tomorrow_short": "TOM",
     "total": "Total",
     "yes": "Yes",
-    "no": "No"
+    "no": "No",
+    "back": "Back"
   },
   "role_selector": {
     "welcome": "Welcome",
