+++ apps/mobile/src/i18n/locales/es.json
@@ -34,7 +34,11 @@
     "whatsapp_placeholder": "3624 000000",
     "first_name_label": "Nombre",
     "last_name_label": "Apellido",
-    "submit_button": "Comenzar"
+    "submit_button": "Comenzar",
+    "email_label": "Email",
+    "password_label": "Contraseña",
+    "admin_portal": "Portal de Administrador",
+    "entrepreneur_panel": "Panel de Emprendedor"
   },
   "validation": {
     "name_required": "El nombre es requerido",
@@ -218,7 +222,8 @@
     "tomorrow_short": "MAÑ",
     "total": "Total",
     "yes": "Sí",
-    "no": "No"
+    "no": "No",
+    "back": "Volver"
   },
   "role_selector": {
     "welcome": "Bienvenido",
